add user Controller and error Handler with some code improvement

+++ src/utils/removeFile.js
@@ -1,15 +1,18 @@
 import fs from "fs";
 
 const removeLocalFiles = ({ files }) => {
-  console.log("internal remove local files")
   if (!files || files.length === 0) return;
 
   Object.values(files)
-    .filter((file) => file[0].path)
+    .filter((file) => file[0]?.path)
     .forEach((file) => {
-      const path = file[0].path;
-      fs.unlinkSync(path);
+      const path = file[0]?.path;
+      if (path && fs.existsSync(path)) {
+        fs.unlinkSync(path);
+      }
     });
+
+  return;
 };
 
 export { removeLocalFiles };
